requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
structlog
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
azure-monitor-opentelemetry-exporter
pydantic-settings
# Add other dependencies here, e.g.:
# sqlalchemy
# pydantic[email]
